guard against sending a sig before a thread::self is stored
Sometimes we need to send a message to a thread that hasn't started yet. Using
genf semantics this is okay. The message queue is usable before the thread
starts. But if the thread needs signals to break from loops we have to guard
against using the pthread self before it is available. Do this by noting the
sig should be sent and sending it to self when the pthread self becomes
available.