make it possible to bind to all rings in a ringset
When binding to all we allocating a single reader ID across all the rings for
convenience. Possible a weird situation could arise where there is no single
reader id that is free on all rings after a lot of attach/detach, however not
anticipating this to be a problem. Normally system will come up and everything
will attach once and stay that way.