readers: don't immediately grab a block when entering a ring
Just set rhead = whead. When we advance a reader we check if we are "entered"
and release the previous only if we are holding one. This will be read 2 and up
after connecting to a ring or restarting.

Problem with entering is that we may encounter skipped blocks and need to move
backwords, invalidating the simple available test of rhead == whead. We can't
reserve a skipped block because then we will be registering a flase skip. So
simplify by not touching any blocks until the first read.