mask the interrupted case in read-wait
The wait_event_interruptible function returns ERESTARTSYS not NOHAND. Catch
this case and return -EINTR (the restarts should not be returned to the user).
Then in the wrapper we catch the EINTR case and return 0, since the function is
designed to return when either there is data or the thread has been
interrupted.