copies of write first/second for multi-writer
Increasingly likely we will clone and specialize for incoming commands as the
requirements create implementation conflicts.

For example multi-writers is easier if we don't need to look at descriptor
bits, just bump a head. Another example, scanning ahead with multiple writers
for a write location will cause incorrect skip computation.

Probably also want to move to an allocate and release model like in genf,
rather than a ring buffer model as in kring because we don't want one writer to
be able to block another. We need to use this for user to kernel commands