using NULL for list termination
The ever-increasing assumption doesn't work because in the time gap between
allocation and list append there is a race between writers. Obvious now, though
not seen before.

Instead will work towards a generation counter for discovering that list
traversals to find the true tail for writing are no longer valid. The
generation can be managed by the readers, along with the head and the maybe
tail.