shifting to free list for alloc and two step tail append
Using a free list which allows a simple free pointer replace as the atomic push
and pop.

Tail append is a two step process. Replace the tail point and if that succees,
add the link, which releases to the reader. Note we can have failure in between
the steps and either need interrpts off or to implement some kind of recovery.