let the thread manage removal from FD list
The thread only sets the closed bit. When the FD list it walked for events we
check this bit and block callbacks of newly closed file descriptors. We don't
alter the list. When building the FD sets on the next iteration we check the
closed bit and detach from the list then.