changed the format of the packet wire data to support simple resend
List of blocks of form ( next-len, content ), where next-len refers to the size
of the next block, including next-len field. The size of the first block is
contained in the packet header, which is at the head of the content. To read,
take in the first next-len and the packet header, then read the remainder of
the first block.