added a packet writer that can send through ITC queues
In many cases we need to hand data off to a service thread and send it out from
there. In the current framework requires unpacking ITC message and repacking as
a packet. Instead use a packet writer to create a packet and send it over an
ITC queue. It can then be forwarded from there.