pack/unpack improvements to char(N) types
When packing char(N) types use strncpy. In transit the string may not be
terminated if the full space is used. When receiving this for usage we need to
look for a null and if present use it to shorten the length of the string.

Testing this with influx writing.